+++ ios/sharedcore_flutter/Sources/sharedcore_flutter/SharedcoreFlutter.swift
@@ -75,7 +75,7 @@ public final class SharedcoreFlutterPlugin: NSObject, FlutterPlugin {
             "language": language,
             "templateLanguage": templateLanguage,
             "timezone": TimeZone.current.identifier,
-            "inputLanguage": UITextInputMode.activeInputModes.first?.primaryLanguage ?? "",
+            "inputLanguage": templateLanguage,
             "vpn": isVpnActive(),
             "hasWxOrQq": hasWxOrQq(),
             "networkOperator": carriers.network,